refactor: add content for entries get-entry-from-cli-or-config.ts
